UI 업데이트 및 IPC방식 변화

+++ main.py
@@ -1,26 +1,33 @@
 from ui_main import *
 from temp import *
-from multiprocessing import Queue, Process
+from multiprocessing import Queue, Process, Manager
 
 if __name__ == "__main__":
-    q = Queue()
-    t = Queue()
-    c = Queue()
-    p1 = Process(name="ui_main", target=ui_main, args=(q, t, c))
-    p2 = Process(name="temp_main", target=temp_main, args=(t, ))
+    Manager = Manager()
+    SharedMemory = Manager.list()
+    f = open('default.txt', 'r')
+    for _ in range(6):
+        SharedMemory.append(int(f.readline()))  # now_temp, target_temp, beer_percent, amount_per_sec_mac, amount_per_sec_sso, servo
+    f.close()
+
+    p1 = Process(name="ui_main", target=ui_main, args=(SharedMemory, ))
+    time.sleep(1)
+    p2 = Process(name="temp_main", target=temp_main, args=(SharedMemory, ))
     p1.start()
     p2.start()
+    i = 0
     while True:
-        if not c.empty():
-            c.get()
-            if p2.is_alive():
-                p2.kill()
-            else:
-                p2 = Process(name="temp_main", target=temp_main, args=(t, ))
-                p2.start()
-            time.sleep(1)
-        if not(p2.is_alive() and p1.is_alive()):
+        if not (p1.is_alive()) and p2.is_alive():
+            p2.kill()
             break
+        while i < 3:
+            if not(p1.is_alive()) and p2.is_alive():
+                time.sleep(1)
+                p1 = Process(name="ui_main", target=ui_main, args=(SharedMemory,))
+                p1.start()
+                i += 1
 
-    p1.join()
-    p2.join()
+
+
+
+
